requirements.txt
# FastAPI and server
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# PDF processing
pdfminer.six>=20221105
PyPDF2>=3.0.0

# LLM
openai>=1.0.0

# Embeddings
sentence-transformers>=2.2.2
# Note: torch will be installed automatically by sentence-transformers
# If you get torch installation errors, see INSTALL.md

# Vector database
chromadb>=0.4.15

# Data handling
pydantic>=2.0.0
pandas>=2.0.0
numpy<2.0.0  # Pin to NumPy 1.x for compatibility with compiled extensions

# Frontend
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
aiofiles>=23.2.1

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
